requirements.txt
alabaster==0.7.13
altgraph==0.17.4
astropy==4.3.1
Babel==2.14.0
cached-property==1.5.2
cachetools==5.3.3
certifi @ file:///C:/b/abs_85o_6fm0se/croot/certifi_1671487778835/work/certifi
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coordinate-conversion==0.1.4.4
cx-Logging==3.1.0
cycler==0.11.0
docutils==0.19
fonttools==4.38.0
future==1.0.0
ghp-import==2.1.0
gitdb==4.0.11
GitPython==3.1.43
griffe==0.30.1
h5py==3.8.0
idna==3.7
imagesize==1.4.1
importlib-metadata==6.7.0
iso8601==2.1.0
Jinja2==3.1.4
jplephem==2.22
kiwisolver==1.4.5
lief==0.12.3
Markdown==3.4.4
markdown-it-py==2.2.0
MarkupSafe==2.1.5
matplotlib==3.5.3
mdit-py-plugins==0.3.5
mdurl==0.1.2
mergedeep==1.3.4
mkdocs==1.5.3
mkdocs-autorefs==0.4.1
mkdocs-git-authors-plugin==0.8.0
mkdocs-git-revision-date-localized-plugin==1.2.2
mkdocs-material==9.2.7
mkdocs-material-extensions==1.2
mkdocstrings==0.22.0
mkdocstrings-python==1.1.2
mpmath==1.3.0
myst-parser==1.0.0
numpy==1.21.6
packaging==24.0
paginate==0.5.6
pandas==1.3.5
pathspec==0.11.2
pefile==2023.2.7
Pillow==9.5.0
platformdirs==4.0.0
pvlib==0.10.4
pyerfa==2.0.0.3
Pygments==2.17.2
pyinstaller==5.13.2
pyinstaller-hooks-contrib==2024.3
pymdown-extensions==10.2.1
PyOpenGL==3.1.7
pyparsing==3.1.2
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pyqtgraph==0.12.4
pyserial==3.5
python-can==4.2.2
python-dateutil==2.9.0.post0
pytz==2024.1
pywin32==306
pywin32-ctypes==0.2.2
PyYAML==6.0.1
pyyaml_env_tag==0.1
regex==2022.10.31
requests==2.31.0
scipy==1.7.3
sgp4==2.23
six==1.16.0
skyfield==1.48
smmap==5.0.1
snowballstemmer==2.2.0
Sphinx==5.3.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sympy==1.10.1
typing_extensions==4.7.1
urllib3==2.0.7
watchdog==3.0.0
wincertstore==0.2
wrapt==1.16.0
zipp==3.15.0
